refactor: moved countdown timer to storybook
