+++ src/stories/forms/QuizForm/QuizForm.css
@@ -3,7 +3,6 @@
   flex-direction: column;
   justify-content: space-evenly;
   text-align: center;
-  min-width: 80%;
   min-height: 200px;
   height: min-content;
   border-radius: 5px;
