+++ src/stories/timer/CountdownTimer/CountdownTimer.css
@@ -1,0 +1,5 @@
+.time-counter {
+  font-size: 100px;
+  color: #61dafb;
+  line-height: 120px;
+}
